split teardown into its own calable method > z_drop.js
